+++ UI_CSharp/IndoorSpace.cs
@@ -10,31 +10,31 @@ namespace midas_challenge
     class IndoorSpace
     {
         [DllImport("Core_CPP.dll", CallingConvention = System.Runtime.InteropServices.CallingConvention.Cdecl)]
-        extern public static IntPtr craeteRoom(double[] coords, IntPtr[] roomList, int n, bool snapmode = true);
+        extern public static IntPtr craeteRoom(int[] coords, IntPtr[] roomList, int n, bool snapmode = true);
 
         [DllImport("Core_CPP.dll", CallingConvention = System.Runtime.InteropServices.CallingConvention.Cdecl)]
-        extern public static IntPtr updateRoom(IntPtr room, double[] coords, IntPtr[] roomList, int n, bool snapmode = true);
+        extern public static IntPtr updateRoom(IntPtr room, int[] coords, IntPtr[] roomList, int n, bool snapmode = true);
         
 
         private Dictionary<int, IntPtr> rooms;
-        public int createRoom(int room_id, List<double> coords)
+        public int createRoom(int room_id, List<int> coords)
         {
 
             return 0;
         }
 
-        public int updateRoom(int room_id, List<double> coords)
+        public int updateRoom(int room_id, List<int> coords)
         {
 
             return 0;
         }
 
-        public int removeRoom(int room_id, List<double> coords)
+        public int removeRoom(int room_id, List<int> coords)
         {
 
         }
 
-        public List<KeyValuePair<int, List<double>>> readRoomList()
+        public List<KeyValuePair<int, List<int>>> readRoomList()
         {
 
         }
